Added Data loader

+++ src/main/java/com/ngrx/course/ngrx_backend/Controller/LessonController.java
@@ -1,8 +1,6 @@
 package com.ngrx.course.ngrx_backend.Controller;
 
-import com.ngrx.course.ngrx_backend.Entity.Course;
 import com.ngrx.course.ngrx_backend.Entity.Lesson;
-import com.ngrx.course.ngrx_backend.Service.CourseService;
 import com.ngrx.course.ngrx_backend.Service.LessonService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -21,7 +19,7 @@ public class LessonController {
     }
 
     @GetMapping("/course/{courseId}")
-    public List<Lesson> getLessonsByCourseId(@PathVariable Long courseId) {
+    public List<Lesson> getLessonsByCourseId(@PathVariable Integer courseId) {
         return lessonService.findByCourseId(courseId);
     }
 
@@ -31,7 +29,7 @@ public class LessonController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteLesson(@PathVariable Long id) {
+    public void deleteLesson(@PathVariable Integer id) {
         lessonService.delete(id);
     }
 }
